Rename resource-show file for SonarCloud
