+++ src/components/resource/resource-show.js
